+++ opentelemetry-stdout/src/logs/transform.rs
@@ -108,17 +108,25 @@ impl From<opentelemetry_sdk::export::logs::LogData> for LogRecord {
     fn from(value: opentelemetry_sdk::export::logs::LogData) -> Self {
         LogRecord {
             attributes: {
-                let mut attributes = value
+                let attributes = value
                     .record
                     .attributes_iter()
                     .map(|(k, v)| KeyValue::from((k.clone(), v.clone()))) // Map each pair to a KeyValue
                     .collect::<Vec<KeyValue>>(); // Collect into a Vec<KeyValue>s
+
+                #[cfg(feature = "populate-logs-event-name")]
                 if let Some(event_name) = &value.record.event_name {
-                    attributes.push(KeyValue::from((
+                    let mut attributes_with_name = attributes;
+                    attributes_with_name.push(KeyValue::from((
                         "name".into(),
                         opentelemetry::Value::String(event_name.clone().into()),
                     )));
+                    attributes_with_name
+                } else {
+                    attributes
                 }
+
+                #[cfg(not(feature = "populate-logs-event-name"))]
                 attributes
             },
             trace_id: value